Make instructions clearer for ParaView newbies
The current instructions are insufficient for those who are new to data visualization software. Especially, the part where one has to press “Apply” at the outset to get the rendering isn't self-evident. (Normally one does that only after making any changes.) Further, removing the lighting results in a black rendering so I removed that.

Tested on ParaView 5.4.1 on Kubuntu Bionic.